requirements.txt
# Core
numpy>=1.24,<3.0
scipy>=1.10,<2.0
pandas>=2.0,<3.0

# PyTorch (install the correct CUDA wheel separately if needed)
torch>=2.1,<3.0

# RL environments
gymnasium>=0.29,<1.0
shimmy>=1.3,<2.0

# MuJoCo (HalfCheetah-v4 etc.)
mujoco>=2.3,<4.0
gymnasium[mujoco]>=0.29,<1.0

# MiniGrid (sparse/exploration tasks)
minigrid>=2.3,<3.0

# Config + CLI
pyyaml>=6.0,<7.0
omegaconf>=2.3,<3.0
hydra-core>=1.3,<2.0

# Logging
tensorboard>=2.13,<3.0
tqdm>=4.66,<5.0

# Plotting + analysis
matplotlib>=3.7,<4.0
seaborn>=0.13,<1.0

# Optional experiment tracking (uncomment if you use it)
# wandb>=0.16,<1.0

# Utilities
rich>=13.6,<14.0
